Create a helper class for the globbing
This class will contain the code required to perform the individual
steps in the glob.

The first part is generating the regex string that will be used in the
pattern matching
